% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\name{AssociationTest}
\alias{AssociationTest}
\title{AssociationTest:}
\usage{
AssociationTest(DataM, gene, peak, method, formula, celltype, res)
}
\arguments{
\item{DataM}{Data Matrix including RNA count, ATAC count and meta data}

\item{method}{Statistic method used for association Test}
}
\value{
res Statistic result
}
\description{
AssociationTest:
}
