% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\name{DataCheck_function}
\alias{DataCheck_function}
\title{DataCheck and Type Checking for CreateOpen4GeneObject Constructor}
\usage{
DataCheck_function(object)
}
\description{
DataCheck and Type Checking for CreateOpen4GeneObject Constructor
}
