% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\name{Extract.gene.peak.pair}
\alias{Extract.gene.peak.pair}
\title{Extract.gene.peak.pair: code for extracting gene~peak pairs based on gene annotation and peak locations}
\usage{
Extract.gene.peak.pair(object)
}
\arguments{
\item{object}{gene.annotation GRanges}
}
\value{
Open4Gene object with updated gene.peak.pair
}
\description{
Extract.gene.peak.pair: code for extracting gene~peak pairs based on gene annotation and peak locations
}
