% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\name{Extract.Peak2Gene.Pairs}
\alias{Extract.Peak2Gene.Pairs}
\title{Extract.Peak2Gene.Pairs: Extracting Peak~Gene Pairs using bedtoolsr}
\usage{
Extract.Peak2Gene.Pairs(object)
}
\arguments{
\item{object}{Gene.Annotation GRanges}
}
\value{
Open4Gene object with Peak2Gene.Pairs extracted
}
\description{
Extract.Peak2Gene.Pairs: Extracting Peak~Gene Pairs using bedtoolsr
}
