% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\name{Open4Gene}
\alias{Open4Gene}
\title{Open4Gene: Main function}
\usage{
Open4Gene(
  object,
  Celltype = "All",
  Binary = FALSE,
  Method = "hurdle",
  MinNum.Cells = 5
)
}
\arguments{
\item{object}{Open4Gene object}

\item{Celltype}{User specified cell type defined in Celltypes column of Meta.data. "All" to test using all cells. "Each" to run test for each cell type. Others, the test will be perform for a given cell type.}

\item{Binary}{If TRUE, the ATAC > 1 will be converted as 1.}

\item{Method}{"hurdle" for Zero-inflated Negative Binomial Regression based on hurdle model. Statistic method used to calculate the correlation between ATAC and RNA.}

\item{MinNum.Cells}{Minimal number of cells with expression (RNA > 0) and open chromatin (ATAC > 0) for association test.}
}
\value{
Open4Gene object with Results from hurdle model
}
\description{
Open4Gene: Main function
}
