% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\name{AssociationTest}
\alias{AssociationTest}
\title{AssociationTest:}
\usage{
AssociationTest(DatMat, Gene, Peak, Method, Formula, Celltype, Res)
}
\arguments{
\item{DatMat}{Data Matrix including RNA count, ATAC count and meta data}

\item{Method}{Statistic Method used for association Test}
}
\value{
Res Statistic Result
}
\description{
AssociationTest:
}
