% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open4Gene.R
\docType{class}
\name{Open4Gene-class}
\alias{Open4Gene-class}
\alias{CreateOpen4GeneObj}
\title{Open4Gene Class Constructor}
\value{
Open4Gene object to use for further analysis
}
\description{
Open4Gene Class Constructor
}
\section{Slots}{

\describe{
\item{\code{RNA}}{dgCMatrix. scRNAseq matrix read as a sparse matrix}

\item{\code{ATAC}}{dgCMatrix. scATACseq matrix read as a sparse matrix}

\item{\code{meta.data}}{data.frame. Metadata table with covariates and a cell ID column ("cell")}

\item{\code{gene.peak.pair}}{data.frame. Dataframe that contains gene-peak pairs for Open4Gene to search through}

\item{\code{gene.peak.dis}}{integer. Distance (peak to gene body) used to extract gene.peak.pair}

\item{\code{gene.annotation}}{GRanges. Gene annotation used to extract gene.peak.pair}

\item{\code{covariates}}{character. Assign covariates that are needed for the analysis. Must be names that are in the columns of meta.data}

\item{\code{celltypes}}{character. Assign celltype column from meta.data}

\item{\code{res}}{data.frame. Table for result of association test, which is initialized as empty.}
}}

